+++ rpc/namespaces/ethereum/eth/api.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/rpc"
 
-	"cosmossdk.io/log"
+	"cosmossdk.io/log/v2"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"